+++ packages/augur-artifacts/src/environments/kovan.json
@@ -80,6 +80,7 @@
     "startWSS": true
   },
   "plugins": {
+    "chat": "3box",
     "comments": "3box"
   }
 }
