+++ packages/augur-artifacts/src/environments/mainnet.json
@@ -65,6 +65,7 @@
     "ZeroXTrade": "0xF6e791393480eD275Cb1729b99C7806EBFc194C0"
   },
   "plugins": {
+    "chat": "3box",
     "comments": "3box"
   }
 }
