+++ packages/augur-ui/src/modules/global-chat/components/global-chat.styles.less
@@ -27,6 +27,7 @@
   div[class*="sc-chat-window"] {
     border: @size-1 solid @color-secondary-action-outline;
     border-radius: @size-3;
+    height: 300px;
     overflow: hidden;
   }
 
@@ -65,6 +66,10 @@
     fill: @color-primary-text;
   }
 
+  div[class*="sc-header--close-button"] svg {
+    padding: @size-4;
+  }
+
   div[class*="sc-header--close-button"] path {
     stroke: @color-primary-text;
   }
@@ -82,99 +87,26 @@
   }
 
   div[class*="sc-message--text"] {
+    .text-12;
+
     background-color: @color-dark-grey !important;
     color: @color-primary-text !important;
+  }
+
+  div[class*="sc-header--team-name"] {
+    padding: 0 @size-8;
   }
 
   div[class*="onlineList_members"] {
     overflow: auto;
   }
 
-  //> div:first-of-type {
-  //  display: none;
-  //}
-  //
-  //> div.ShowGlobalChat {
-  //  display: block;
-  //  line-height: 0; // This removes the iframe's extra padding
-  //  border: @size-1 solid @color-secondary-action-outline;
-  //  border-radius: @size-3;
-  //  overflow: hidden;
-  //}
+  div[class*="sc-header"] {
+    padding: 0 @size-10;
+    min-height: unset;
+  }
 
-  // Closed global chat window
-  //> button {
-  //  .mono-9-bold;
-  //
-  //  align-items: center;
-  //  background: @color-module-background;
-  //  color: @color-primary-text;
-  //  display: flex;
-  //  flex-direction: row;
-  //  text-transform: uppercase;
-  //
-  //  > svg {
-  //    fill: @color-primary-text;
-  //    height: @size-12;
-  //    margin-left: @size-8;
-  //    transform: rotate(180deg);
-  //    width: @size-12;
-  //  }
-  //}
-
-  // Open global chat window
-  //> div {
-  //  background-color: @color-module-background;
-  //
-  //  // Header section
-  //  > div:first-of-type {
-  //    align-items: center;
-  //    background-color: @color-dark-grey;
-  //    border-bottom: 1px solid @color-dark-grey;
-  //    display: flex;
-  //    flex-direction: row;
-  //    padding: @size-7 @size-8;
-  //
-  //    > span {
-  //      color: @color-primary-text;
-  //      margin: 0 0;
-  //      text-transform: uppercase;
-  //    }
-  //
-  //    > span:first-of-type {
-  //      .mono-12;
-  //
-  //      margin: 0 @size-8 0 0;
-  //      font-weight: @weight-bold;
-  //    }
-  //
-  //    > span:nth-of-type(2) {
-  //      .mono-9;
-  //
-  //      background: fade(@color-open, 15%);
-  //      border-radius: 0.1875rem;
-  //      color: @color-open;
-  //      padding: 0 1px;
-  //    }
-  //
-  //    > button {
-  //      margin: 0 0 0 auto;
-  //
-  //      > svg {
-  //        height: @size-8;
-  //        fill: @color-primary-text;
-  //        margin: 0;
-  //        width: @size-8;
-  //        position: relative;
-  //        bottom: 2px;
-  //      }
-  //    }
-  //  }
-  //
-  //  iframe {
-  //    border: 0px;
-  //    height: 16rem;
-  //    width: 18.75rem;
-  //  }
-  //}
+  h3[class*="sc-header--team-name_text"] {
+    .text-16-bold;
+  }
 }
