+++ packages/augur-ui/src/modules/market/components/common/comments/market-comments.tsx
@@ -1,12 +1,11 @@
-import React, { useEffect, useState } from 'react';
-import Box from '3box';
+import React from 'react';
 import ThreeBoxComments from '3box-comments-react';
 import { FacebookComments } from 'modules/market/components/common/comments/facebook-comments';
 
 import Styles from 'modules/market/components/market-view/market-view.styles.less';
+import { use3box } from 'utils/use-3box';
 
 interface MarketCommentsProps {
-  accountType: string;
   adminEthAddr: string;
   colorScheme: string;
   marketId: string;
@@ -17,7 +16,6 @@ interface MarketCommentsProps {
 }
 
 export const MarketComments = ({
-  accountType,
   adminEthAddr,
   colorScheme,
   marketId,
@@ -26,65 +24,23 @@ export const MarketComments = ({
   provider,
   whichCommentPlugin,
 }: MarketCommentsProps) => {
-  const [address, setAddress] = useState();
-  const [box, setBox] = useState({});
-  const [isReady, setIsReady] = useState(false);
-  const [profile, setProfile] = useState();
-
-  useEffect(() => {
-    whichCommentPlugin === '3box' && handleLogin();
-  }, []);
-
-  useEffect(() => {
-    whichCommentPlugin === '3box' && handleLogin();
-  }, [accountType, provider]);
-
-  const handleLogin = async () => {
-    setIsReady(false);
-
-    if (!provider) {
-      return;
-    }
-
-    let threeBoxInstance;
-    let addressFromProvider = (await provider.enable())[0];
-    let publicProfile;
-
-    try {
-      threeBoxInstance = await Box.create(provider);
-
-      await threeBoxInstance.auth([], { address: addressFromProvider });
-      await threeBoxInstance.syncDone;
-
-      const space = await threeBoxInstance.openSpace('augur', {});
-      await space.syncDone;
-
-      publicProfile = await Box.getProfile(addressFromProvider);
-    } catch (error) {
-      console.error(error);
-      return;
-    }
-
-    setBox(threeBoxInstance);
-    setAddress(addressFromProvider);
-    setProfile(publicProfile);
-    setIsReady(true);
-  };
+  const { address, box, isReady, profile } =
+    whichCommentPlugin === '3box' && use3box(provider);
 
   return (
     <section className={Styles.Comments}>
-      {whichCommentPlugin === '3box' && isReady && (
+      {whichCommentPlugin === '3box' && threeBox.isReady && (
         <ThreeBoxComments
           // required
           spaceName="augur"
           threadName={marketId}
           adminEthAddr={adminEthAddr}
           // Required props for context A) & B)
-          box={box}
-          currentUserAddr={address}
+          box={threeBox.box}
+          currentUserAddr={threeBox.address}
           // optional
           showCommentCount={numPosts}
-          currentUser3BoxProfile={profile}
+          currentUser3BoxProfile={threeBox.profile}
           // useHovers={true}
         />
       )}
